services/user-service/requirements.txt
Flask
psycopg2-binary
flask-sqlalchemy
python-dotenv
PyJWT
